x/poc/types/message_deregister_app_test.go
package types

import (
	"testing"

	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
	"github.com/stretchr/testify/require"
	"poc/testutil/sample"
)

func TestMsgDeregisterApp_ValidateBasic(t *testing.T) {
	tests := []struct {
		name string
		msg  MsgDeregisterApp
		err  error
	}{
		{
			name: "invalid address",
			msg: MsgDeregisterApp{
				Creator: "invalid_address",
			},
			err: sdkerrors.ErrInvalidAddress,
		}, {
			name: "valid address",
			msg: MsgDeregisterApp{
				Creator: sample.AccAddress(),
			},
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			err := tt.msg.ValidateBasic()
			if tt.err != nil {
				require.ErrorIs(t, err, tt.err)
				return
			}
			require.NoError(t, err)
		})
	}
}
